% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{length_of_intervention}
\alias{length_of_intervention}
\title{Total length of a specific intervention.}
\usage{
length_of_intervention(enc_info_df, interventions_df, intervention_name,
  time_units)
}
\arguments{
\item{enc_info_df}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   }}

\item{interventions_df}{\code{data frame}: provides intervention events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{POSIXct}: Start.Time
   \item \code{POSIXct}: End.Time
   }}

\item{intervention_name}{\code{chr}: names the intervention of interest}

\item{time_units}{\code{chr}: time units to compute time difference in}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: length of intervention_name
   }
}
\description{
\code{length_of_intervention} returns the total length of the specified
intervention.
}
\details{
Computes the total length of intervention, per encounter. Encounters with
no events of the specified interventions are assumed to not have had the
intervention, and a value of zero is given. The function currently works
only for interventions that have both a start time and end time.
}

