% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{on_intervention}
\alias{on_intervention}
\title{Intervention performed indicator.}
\usage{
on_intervention(enc_info_df, interventions_df, intervention_name)
}
\arguments{
\item{enc_info_df}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   }}

\item{interventions_df}{\code{data frame}: provides intervention events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   }}

\item{intervention_name}{\code{chr}: names the intervention
of interest}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: intervention_name
   }
}
\description{
\code{on_intervention} returns an indication of whether a particular
intervention was performed during an encounter, for every encounter.
}
\details{
Determines whether a particular intervention was performed at any point in
time, per encounter. A column in the resulting data frame indicates whether
the intervention was performed (with a 1) or not (with a 0).
}

