% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{oxygen_saturation_index}
\alias{oxygen_saturation_index}
\title{Instantaneous Oxygen Saturation Index (OSI).}
\usage{
oxygen_saturation_index(interventions, vitals)
}
\arguments{
\item{interventions}{\code{data frame}: provides interventions events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Value
   \item \code{POSIXct}: Start.Time
   }}

\item{vitals}{\code{data frame}: provides vitals events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Value
   \item \code{POSIXct}: End.Time
   }}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: oxygen_saturation_index
   \item \code{POSIXct}: time
   }
}
\description{
\code{oxygen_saturation_index} returns the instantaneous oxygen saturation
index, for each encounter.
}
\details{
As per Nelson Sanchez-Pinto,
OSI = 100 * Mean Airway Pressure * FiO2 / Pulse Oximetry,
where Meand Ariway Pressure and FiO2 are taken to be the most recent within
the last hour preceding Pulse Oximetry. Therefore, each returned OSI
will have the time correpsonding to the Pulse Oximetry measurement.
}

