% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{time_off_intervention}
\alias{time_off_intervention}
\title{Time off a specific intervention, within the first 28 days.}
\usage{
time_off_intervention(enc_info_df, interventions_df, intervention_name,
  time_units)
}
\arguments{
\item{enc_info_df}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{POSIXct}: Episode.Start.Timestamp
   }}

\item{interventions_df}{\code{data frame}: provides intervention events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{POSIXct}: Start.Time
   \item \code{POSIXct}: End.Time
   }}

\item{intervention_name}{\code{chr}: names the intervention of interest}

\item{time_units}{\code{chr}: time units to compute time difference in}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: time off intervention_name
   }
}
\description{
\code{time_off_intervention} returns the time an encounter has been off the
specified intervention within the first 28 days.
}
\details{
Computes time time off since the last event of the specified intervention,
in the first 28 days, per encounter. Missing values are imputed with NAs,
as encounters with no events of the specified interventions are assumed
to not have had that intervention. The function currently works
only for interventions that have both a start time and end time.
}

