% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{worst_gcs}
\alias{worst_gcs}
\title{Worst Glasgow Coma Score.}
\usage{
worst_gcs(enc_info, vitals)
}
\arguments{
\item{enc_info}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   }}

\item{vitals}{\code{data frame}: provides events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Value
   }}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: worst_gcs
   }
}
\description{
\code{worst_gcs} returns the worst (lowest) Glasgow Coma Score (GCS)
for each encounter selected.
}
\details{
Glasgow coma score ranges from 3 to 15. 3 being the worst score and 15 the
best.
Return value will be NA for encounters with no observations of GCS.
}

