% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{worst_inr}
\alias{worst_inr}
\title{Worst INR.}
\usage{
worst_inr(enc_info, labs)
}
\arguments{
\item{labs}{\code{data frame}: provides events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Value
   }}

\item{enc_info_df}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   }}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: worst_inr
   }
}
\description{
\code{worst_inr} returns the worst (highest) INR
for each encounter selected.
}
\details{
As per Nelson Sanchez-Pinto,
worst INR = highest INR
Return value will be 1 for encounters with no observations of INR (per
Nelson Sanchez-Pinto).
}

