% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{worst_is}
\alias{worst_is}
\title{Worst Inotrope Score.}
\usage{
worst_is(enc_info_df, drugs_df)
}
\arguments{
\item{enc_info_df}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{POSIXct}: Episode.Start.Timestamp
   }}

\item{drugs_df}{\code{data frame}: provides events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Rate.Value
   \item \code{POSIXct}: Time.of.Administration
   }}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: worst_is
   }
}
\description{
\code{worst_is} returns the worst (highest) inotrope score, averaged
per hour, for each encounter selected.
}
\details{
As per Nelson Sanchez-Pinto,
Worst Inotrope Score (IS) = highest IS
IS = dopamine + dobutamine + (epinephrine + norepinephrine) * 100 +
milrinone * 10.
The worst inotrope score is selected from among the hourly-averaged scores for
each encounter.
Encounters with no inotrope score are given 0 as their inotrope score,
as per Nelson Sanchez-Pinto.
}

