% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_metrics.R
\name{worst_osi}
\alias{worst_osi}
\title{Worst Oxygen Saturation Index}
\usage{
worst_osi(enc_info, interventions, vitals)
}
\arguments{
\item{interventions}{\code{data frame}: provides interventions events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Value
   \item \code{POSIXct}: Start.Time
   }}

\item{vitals}{\code{data frame}: provides vitals events
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{chr}: Variable.Name
   \item \code{num}: Value
   \item \code{POSIXct}: End.Time
   }}

\item{enc_info_df}{\code{data frame}: provides the encounters of interest
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{POSIXct}: Episode.Start.Timestamp
   }}
}
\value{
data_frame:
\itemize{
   \item \code{chr}: patient_id_encounter
   \item \code{num}: worst_osi
   }
}
\description{
\code{worst_osi} returns the worst (highest) oxygen saturation
index, for each encounter selected.
}
\details{
Missing values are imputed with zero: Encounters that do not have all
measurements (i.e, mean airway pressure, fio2 or pulseox) get OSI = 0,
as per Nelson.
NOTE: This function may take up to 5 minutes on the entire DS 2.1.1.
}

